Extract peak broadening into a domain-neutral $lib/lineshape
broaden_peaks lived in the XRD module, so the IR/Raman code reached across
into $lib/xrd/broadening for it despite sharing none of its Bragg-angle
assumptions. Move the stick-to-continuum convolution to its own module that
assumes neither scale: XRD drives it with Caglioti widths in degrees 2theta,
IR/Raman with constant or frequency-dependent widths in cm^-1.

Exported as the matterviz/lineshape subpath.
